+++ api/server.go
@@ -446,6 +446,7 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 	protoapi.RegisterTradingDataServiceServer(g.srv, g.tradingDataService)
 
 	tradingDataSvcV2 := &tradingDataServiceV2{
+		config:               g.Config,
 		log:                  g.log,
 		v2ApiEnabled:         g.useSQLStores,
 		orderService:         g.orderServiceV2,
@@ -461,7 +462,9 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 		riskService:          g.riskServiceV2,
 		positionService:      g.positionServiceV2,
 		accountService:       g.accountServiceV2,
+		rewardService:        g.rewardServiceV2,
 	}
+
 	protoapi2.RegisterTradingDataServiceServer(g.srv, tradingDataSvcV2)
 
 	eg, ctx := errgroup.WithContext(ctx)